refactor(errors): Split errors into modules
- Moved error-related classes and functions into separate files for better organization.
- Introduced `ClerkAPIResponseError`, `EmailLinkError`, `ClerkRuntimeError`, and `ClerkWebAuthnError` classes.
- Added helper functions for error checking and parsing in `helpers.ts` and `parseError.ts`.
- Deprecated unused error checking functions and streamlined error handling logic.
